refactor(dot): consolidate the dot reporters
