+++ src/report/dot/common/boringModuleColorScheme.json
