+++ src/report/dot/common/coloring.js
