+++ src/report/dot/common/compareOnSource.js
@@ -1,0 +1,1 @@
+module.exports = (pModuleOne, pModuleTwo) => pModuleOne.source > pModuleTwo.source ? 1 : -1;
