+++ src/report/dot/common/richModuleColorScheme.json
