+++ src/report/dot/folderLevel/index.js
@@ -1,44 +1,13 @@
-const path       = require('path').posix;
-const Handlebars = require("handlebars/runtime");
-const _clone     = require('lodash/clone');
-const _get       = require('lodash/get');
-const _uniqBy    = require('lodash/uniqBy');
-const coloring   = require('../dot/coloring');
+const path            = require('path').posix;
+const Handlebars      = require("handlebars/runtime");
+const _clone          = require('lodash/clone');
+const _get            = require('lodash/get');
+const _uniqBy         = require('lodash/uniqBy');
+const coloring        = require('../common/coloring');
+const folderify       = require('../common/folderify');
+const compareOnSource = require("../common/compareOnSource");
 
 require("./ddot.template");
-
-function compareOnSource(pModuleOne, pModuleTwo) {
-    return pModuleOne.source > pModuleTwo.source ? 1 : -1;
-}
-
-function toFullPath (pAll, pCurrent) {
-    return `${pAll}${pCurrent}${path.sep}`;
-}
-
-function aggregate (pPathSnippet, pCounter, pPathArray){
-    return {
-        snippet: pPathSnippet,
-        aggregateSnippet: `${pPathArray.slice(0, pCounter).reduce(toFullPath, '')}${pPathSnippet}`
-    };
-}
-
-function folderify(pModule) {
-    let lAdditions = {};
-    let lDirName = path.dirname(pModule.source);
-
-    if (lDirName !== ".") {
-        lAdditions.folder = lDirName;
-        lAdditions.path   = lDirName.split(path.sep).map(aggregate);
-    }
-
-    lAdditions.label = path.basename(pModule.source);
-
-    return Object.assign(
-        {},
-        pModule,
-        lAdditions
-    );
-}
 
 function colorize(pModule) {
     return Object.assign(
