+++ src/report/dot/moduleLevel/index.js
@@ -1,41 +1,10 @@
-const path       = require('path').posix;
-const Handlebars = require("handlebars/runtime");
-const coloring   = require('./coloring');
+const path            = require('path').posix;
+const Handlebars      = require("handlebars/runtime");
+const coloring        = require('../common/coloring');
+const folderify       = require('../common/folderify');
+const compareOnSource = require("../common/compareOnSource");
 
 require("./dot.template");
-
-function compareOnSource(pModuleOne, pModuleTwo) {
-    return pModuleOne.source > pModuleTwo.source ? 1 : -1;
-}
-
-function toFullPath (pAll, pCurrent) {
-    return `${pAll}${path.sep}${pCurrent}`;
-}
-
-function aggregate (pPathSnippet, pCounter, pPathArray){
-    return {
-        snippet: pPathSnippet,
-        aggregateSnippet: `${pPathArray.slice(0, pCounter).reduce(toFullPath, '')}${path.sep}${pPathSnippet}`
-    };
-}
-
-function folderify(pModule) {
-    let lAdditions = {};
-    let lDirName = path.dirname(pModule.source);
-
-    if (lDirName !== ".") {
-        lAdditions.folder = lDirName;
-        lAdditions.path   = lDirName.split(path.sep).map(aggregate);
-    }
-
-    lAdditions.label = path.basename(pModule.source);
-
-    return Object.assign(
-        {},
-        pModule,
-        lAdditions
-    );
-}
 
 function colorize(pColoringScheme) {
     return pModule => Object.assign(
