+++ test/cli/fixtures/cjs.dir.filtered.dot
@@ -15,15 +15,15 @@ digraph "dependency-cruiser output"{
     "fs" [label="fs" color="grey" fontcolor="grey" ]
     "http" [label="http" color="grey" fontcolor="grey" ]
     "path" [label="path" color="grey" fontcolor="grey" ]
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/one_only_one.js" [label="one_only_one.js" URL="test/cli/fixtures/cjs/one_only_one.js"] } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/one_only_two.js" [label="one_only_two.js" URL="test/cli/fixtures/cjs/one_only_two.js"] } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/root_one.js" [label="root_one.js" URL="test/cli/fixtures/cjs/root_one.js"] } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/root_two.js" [label="root_two.js" URL="test/cli/fixtures/cjs/root_two.js"] } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/shared.js" [label="shared.js" URL="test/cli/fixtures/cjs/shared.js"] } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/somedata.json" [label="somedata.json" fillcolor="#ffee44" URL="test/cli/fixtures/cjs/somedata.json"] } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" subgraph "cluster_/test/cli/fixtures/cjs/sub" {label="sub" "test/cli/fixtures/cjs/sub/depindir.js" [label="depindir.js" URL="test/cli/fixtures/cjs/sub/depindir.js"] } } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" subgraph "cluster_/test/cli/fixtures/cjs/sub" {label="sub" "test/cli/fixtures/cjs/sub/dir.js" [label="dir.js" URL="test/cli/fixtures/cjs/sub/dir.js"] } } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/two_only_one.js" [label="two_only_one.js" URL="test/cli/fixtures/cjs/two_only_one.js"] } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/one_only_one.js" [label="one_only_one.js" URL="test/cli/fixtures/cjs/one_only_one.js"] } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/one_only_two.js" [label="one_only_two.js" URL="test/cli/fixtures/cjs/one_only_two.js"] } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/root_one.js" [label="root_one.js" URL="test/cli/fixtures/cjs/root_one.js"] } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/root_two.js" [label="root_two.js" URL="test/cli/fixtures/cjs/root_two.js"] } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/shared.js" [label="shared.js" URL="test/cli/fixtures/cjs/shared.js"] } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/somedata.json" [label="somedata.json" fillcolor="#ffee44" URL="test/cli/fixtures/cjs/somedata.json"] } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" subgraph "cluster_test/cli/fixtures/cjs/sub" {label="sub" "test/cli/fixtures/cjs/sub/depindir.js" [label="depindir.js" URL="test/cli/fixtures/cjs/sub/depindir.js"] } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" subgraph "cluster_test/cli/fixtures/cjs/sub" {label="sub" "test/cli/fixtures/cjs/sub/dir.js" [label="dir.js" URL="test/cli/fixtures/cjs/sub/dir.js"] } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/cjs" {label="cjs" "test/cli/fixtures/cjs/two_only_one.js" [label="two_only_one.js" URL="test/cli/fixtures/cjs/two_only_one.js"] } } } }
 
     "test/cli/fixtures/cjs/one_only_one.js" -> "path" [color="grey"]
     "test/cli/fixtures/cjs/one_only_two.js" -> "path" [color="grey"]
