+++ test/cli/fixtures/duplicate-subs.dot
@@ -12,12 +12,12 @@ digraph "dependency-cruiser output"{
     node [shape=box style="rounded, filled" fillcolor="#ffffcc" height=0.2 fontname=Helvetica fontsize=9]
     edge [color=black arrowhead=normal fontname=Helvetica fontsize=9]
 
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_/test/cli/fixtures/duplicate-subs/nested-sub" {label="nested-sub" subgraph "cluster_/test/cli/fixtures/duplicate-subs/nested-sub/sub" {label="sub" "test/cli/fixtures/duplicate-subs/nested-sub/sub/index.js" [label="index.js" URL="test/cli/fixtures/duplicate-subs/nested-sub/sub/index.js"] } } } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_/test/cli/fixtures/duplicate-subs/nested-sub" {label="nested-sub" subgraph "cluster_/test/cli/fixtures/duplicate-subs/nested-sub/sub" {label="sub" "test/cli/fixtures/duplicate-subs/nested-sub/sub/sub.js" [label="sub.js" URL="test/cli/fixtures/duplicate-subs/nested-sub/sub/sub.js"] } } } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_/test/cli/fixtures/duplicate-subs/nested-sub" {label="nested-sub" subgraph "cluster_/test/cli/fixtures/duplicate-subs/nested-sub/sub" {label="sub" "test/cli/fixtures/duplicate-subs/nested-sub/sub/wakka.js" [label="wakka.js" URL="test/cli/fixtures/duplicate-subs/nested-sub/sub/wakka.js"] } } } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_/test/cli/fixtures/duplicate-subs/sub" {label="sub" "test/cli/fixtures/duplicate-subs/sub/index.js" [label="index.js" URL="test/cli/fixtures/duplicate-subs/sub/index.js"] } } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_/test/cli/fixtures/duplicate-subs/sub" {label="sub" "test/cli/fixtures/duplicate-subs/sub/more-in-sub.js" [label="more-in-sub.js" URL="test/cli/fixtures/duplicate-subs/sub/more-in-sub.js"] } } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_/test/cli/fixtures/duplicate-subs/sub" {label="sub" "test/cli/fixtures/duplicate-subs/sub/most-in-sub.js" [label="most-in-sub.js" URL="test/cli/fixtures/duplicate-subs/sub/most-in-sub.js"] } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_test/cli/fixtures/duplicate-subs/nested-sub" {label="nested-sub" subgraph "cluster_test/cli/fixtures/duplicate-subs/nested-sub/sub" {label="sub" "test/cli/fixtures/duplicate-subs/nested-sub/sub/index.js" [label="index.js" URL="test/cli/fixtures/duplicate-subs/nested-sub/sub/index.js"] } } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_test/cli/fixtures/duplicate-subs/nested-sub" {label="nested-sub" subgraph "cluster_test/cli/fixtures/duplicate-subs/nested-sub/sub" {label="sub" "test/cli/fixtures/duplicate-subs/nested-sub/sub/sub.js" [label="sub.js" URL="test/cli/fixtures/duplicate-subs/nested-sub/sub/sub.js"] } } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_test/cli/fixtures/duplicate-subs/nested-sub" {label="nested-sub" subgraph "cluster_test/cli/fixtures/duplicate-subs/nested-sub/sub" {label="sub" "test/cli/fixtures/duplicate-subs/nested-sub/sub/wakka.js" [label="wakka.js" URL="test/cli/fixtures/duplicate-subs/nested-sub/sub/wakka.js"] } } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_test/cli/fixtures/duplicate-subs/sub" {label="sub" "test/cli/fixtures/duplicate-subs/sub/index.js" [label="index.js" URL="test/cli/fixtures/duplicate-subs/sub/index.js"] } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_test/cli/fixtures/duplicate-subs/sub" {label="sub" "test/cli/fixtures/duplicate-subs/sub/more-in-sub.js" [label="more-in-sub.js" URL="test/cli/fixtures/duplicate-subs/sub/more-in-sub.js"] } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/duplicate-subs" {label="duplicate-subs" subgraph "cluster_test/cli/fixtures/duplicate-subs/sub" {label="sub" "test/cli/fixtures/duplicate-subs/sub/most-in-sub.js" [label="most-in-sub.js" URL="test/cli/fixtures/duplicate-subs/sub/most-in-sub.js"] } } } } }
 
     "test/cli/fixtures/duplicate-subs/nested-sub/sub/index.js" -> "test/cli/fixtures/duplicate-subs/nested-sub/sub/sub.js"
     "test/cli/fixtures/duplicate-subs/nested-sub/sub/index.js" -> "test/cli/fixtures/duplicate-subs/nested-sub/sub/wakka.js"
