+++ test/report/dot/folderLevel/ddot.spec.js
@@ -1,13 +1,13 @@
 const fs      = require('fs');
 const expect  = require('chai').expect;
-const render  = require('../../../src/report/ddot');
+const render  = require('../../../../src/report/dot/folderLevel');
 const deps    = require('./fixtures/dependency-cruiser-2019-01-14.json');
 const orphans = require('./fixtures/orphans.json');
 const rxjs    = require('./fixtures/rxjs.json');
 
-const consolidatedDot        = fs.readFileSync('test/report/ddot/fixtures/dependency-cruiser-2019-01-14.dot', 'utf8');
-const consolidatedOrphansDot = fs.readFileSync('test/report/ddot/fixtures/consolidated-orphans.dot', 'utf8');
-const consolidatedRxJs       = fs.readFileSync('test/report/ddot/fixtures/consolidated-rxjs.dot', 'utf8');
+const consolidatedDot        = fs.readFileSync('test/report/dot/folderLevel/fixtures/dependency-cruiser-2019-01-14.dot', 'utf8');
+const consolidatedOrphansDot = fs.readFileSync('test/report/dot/folderLevel/fixtures/consolidated-orphans.dot', 'utf8');
+const consolidatedRxJs       = fs.readFileSync('test/report/dot/folderLevel/fixtures/consolidated-rxjs.dot', 'utf8');
 
 describe("ddot reporter", () => {
     it("consolidates to folder level", () => {
