+++ test/report/dot/folderLevel/fixtures/consolidated-rxjs.dot
