+++ test/report/dot/folderLevel/fixtures/dependency-cruiser-2019-01-14.json
