+++ test/report/dot/moduleLevel/rcdot.spec.js
@@ -1,10 +1,11 @@
-const fs          = require('fs');
-const expect      = require('chai').expect;
-const renderRCDot = require('../../../src/report/dot')(require('../../../src/report/dot/richModuleColorScheme.json'));
-const renderDot   = require('../../../src/report/dot')();
-const renderBoringDot =
-    require('../../../src/report/dot')(require('../../../src/report/dot/boringModuleColorScheme.json'));
-const deps        = require('../fixtures/richmodulecolor.json');
+const fs              = require('fs');
+const expect          = require('chai').expect;
+const RCScheme        = require('../../../../src/report/dot/common/richModuleColorScheme.json');
+const renderRCDot     = require('../../../../src/report/dot/moduleLevel')(RCScheme);
+const renderDot       = require('../../../../src/report/dot/moduleLevel')();
+const boringScheme    = require('../../../../src/report/dot/common/boringModuleColorScheme.json');
+const renderBoringDot = require('../../../../src/report/dot/moduleLevel')(boringScheme);
+const deps            = require('../../fixtures/richmodulecolor.json');
 
 const richColorFixture = fs.readFileSync('test/report/fixtures/richmodulecolor.dot', 'utf8');
 const boringColorFixture = fs.readFileSync('test/report/fixtures/defaultmodulecolor.dot', 'utf8');
