+++ test/report/fixtures/clusterless.dot
@@ -14,9 +14,9 @@ digraph "dependency-cruiser output"{
 
     "fs" [label="fs" color="grey" fontcolor="grey" ]
     "http" [label="http" color="grey" fontcolor="grey" ]
-    subgraph "cluster_/node_modules" {label="node_modules" subgraph "cluster_/node_modules/somemodule" {label="somemodule" subgraph "cluster_/node_modules/somemodule/node_modules" {label="node_modules" subgraph "cluster_/node_modules/somemodule/node_modules/someothermodule" {label="someothermodule" "node_modules/somemodule/node_modules/someothermodule/main.js" [label="main.js" URL="node_modules/somemodule/node_modules/someothermodule/main.js"] } } } }
-    subgraph "cluster_/node_modules" {label="node_modules" subgraph "cluster_/node_modules/somemodule" {label="somemodule" subgraph "cluster_/node_modules/somemodule/src" {label="src" "node_modules/somemodule/src/moar-javascript.js" [label="moar-javascript.js" URL="node_modules/somemodule/src/moar-javascript.js"] } } }
-    subgraph "cluster_/node_modules" {label="node_modules" subgraph "cluster_/node_modules/somemodule" {label="somemodule" subgraph "cluster_/node_modules/somemodule/src" {label="src" "node_modules/somemodule/src/somemodule.js" [label="somemodule.js" URL="node_modules/somemodule/src/somemodule.js"] } } }
+    subgraph "cluster_node_modules" {label="node_modules" subgraph "cluster_node_modules/somemodule" {label="somemodule" subgraph "cluster_node_modules/somemodule/node_modules" {label="node_modules" subgraph "cluster_node_modules/somemodule/node_modules/someothermodule" {label="someothermodule" "node_modules/somemodule/node_modules/someothermodule/main.js" [label="main.js" URL="node_modules/somemodule/node_modules/someothermodule/main.js"] } } } }
+    subgraph "cluster_node_modules" {label="node_modules" subgraph "cluster_node_modules/somemodule" {label="somemodule" subgraph "cluster_node_modules/somemodule/src" {label="src" "node_modules/somemodule/src/moar-javascript.js" [label="moar-javascript.js" URL="node_modules/somemodule/src/moar-javascript.js"] } } }
+    subgraph "cluster_node_modules" {label="node_modules" subgraph "cluster_node_modules/somemodule" {label="somemodule" subgraph "cluster_node_modules/somemodule/src" {label="src" "node_modules/somemodule/src/somemodule.js" [label="somemodule.js" URL="node_modules/somemodule/src/somemodule.js"] } } }
     "one_only_one.js" [label="one_only_one.js" URL="one_only_one.js"]
     "one_only_two.js" [label="one_only_two.js" URL="one_only_two.js"]
     "path" [label="path" color="grey" fontcolor="grey" ]
@@ -24,8 +24,8 @@ digraph "dependency-cruiser output"{
     "root_two.js" [label="root_two.js" URL="root_two.js"]
     "shared.js" [label="shared.js" color="red" fontcolor="red" ]
     "somedata.json" [label="somedata.json" fillcolor="#ffee44" URL="somedata.json"]
-    subgraph "cluster_/sub" {label="sub" "sub/depindir.js" [label="depindir.js" URL="sub/depindir.js"] }
-    subgraph "cluster_/sub" {label="sub" "sub/dir.js" [label="dir.js" URL="sub/dir.js"] }
+    subgraph "cluster_sub" {label="sub" "sub/depindir.js" [label="depindir.js" URL="sub/depindir.js"] }
+    subgraph "cluster_sub" {label="sub" "sub/dir.js" [label="dir.js" URL="sub/dir.js"] }
     "two_only_one.js" [label="two_only_one.js" URL="two_only_one.js"]
 
     "node_modules/somemodule/src/somemodule.js" -> "node_modules/somemodule/src/moar-javascript.js" [color="orange" penwidth=2.0 tooltip="unnamed" fontcolor="orange" xlabel="unnamed"]
