+++ test/report/fixtures/defaultmodulecolor.dot
@@ -12,15 +12,15 @@ digraph "dependency-cruiser output"{
     node [shape=box style="rounded, filled" fillcolor="#ffffcc" height=0.2 fontname=Helvetica fontsize=9]
     edge [color=black arrowhead=normal fontname=Helvetica fontsize=9]
 
-    subgraph "cluster_/src" {label="src" "src/coffee.coffee" [label="coffee.coffee" URL="src/coffee.coffee"] }
-    subgraph "cluster_/src" {label="src" "src/javascript.js" [label="javascript.js" URL="src/javascript.js"] }
-    subgraph "cluster_/src" {label="src" "src/json.json" [label="json.json" fillcolor="#ffee44" URL="src/json.json"] }
-    subgraph "cluster_/src" {label="src" "src/litcoffee.litcoffee" [label="litcoffee.litcoffee" URL="src/litcoffee.litcoffee"] }
-    subgraph "cluster_/src" {label="src" "src/litcoffeemarkdown.coffee.md" [label="litcoffeemarkdown.coffee.md" URL="src/litcoffeemarkdown.coffee.md"] }
-    subgraph "cluster_/src" {label="src" "src/reactthing.cjsx" [label="reactthing.cjsx" URL="src/reactthing.cjsx"] }
-    subgraph "cluster_/src" {label="src" "src/reactthing.csx" [label="reactthing.csx" URL="src/reactthing.csx"] }
-    subgraph "cluster_/src" {label="src" "src/reactthing.jsx" [label="reactthing.jsx" URL="src/reactthing.jsx"] }
-    subgraph "cluster_/src" {label="src" "src/reactthing.tsx" [label="reactthing.tsx" URL="src/reactthing.tsx"] }
-    subgraph "cluster_/src" {label="src" "src/typescript.ts" [label="typescript.ts" URL="src/typescript.ts"] }
+    subgraph "cluster_src" {label="src" "src/coffee.coffee" [label="coffee.coffee" URL="src/coffee.coffee"] }
+    subgraph "cluster_src" {label="src" "src/javascript.js" [label="javascript.js" URL="src/javascript.js"] }
+    subgraph "cluster_src" {label="src" "src/json.json" [label="json.json" fillcolor="#ffee44" URL="src/json.json"] }
+    subgraph "cluster_src" {label="src" "src/litcoffee.litcoffee" [label="litcoffee.litcoffee" URL="src/litcoffee.litcoffee"] }
+    subgraph "cluster_src" {label="src" "src/litcoffeemarkdown.coffee.md" [label="litcoffeemarkdown.coffee.md" URL="src/litcoffeemarkdown.coffee.md"] }
+    subgraph "cluster_src" {label="src" "src/reactthing.cjsx" [label="reactthing.cjsx" URL="src/reactthing.cjsx"] }
+    subgraph "cluster_src" {label="src" "src/reactthing.csx" [label="reactthing.csx" URL="src/reactthing.csx"] }
+    subgraph "cluster_src" {label="src" "src/reactthing.jsx" [label="reactthing.jsx" URL="src/reactthing.jsx"] }
+    subgraph "cluster_src" {label="src" "src/reactthing.tsx" [label="reactthing.tsx" URL="src/reactthing.tsx"] }
+    subgraph "cluster_src" {label="src" "src/typescript.ts" [label="typescript.ts" URL="src/typescript.ts"] }
 
 }
