+++ test/report/fixtures/unresolvable.dot
@@ -13,8 +13,8 @@ digraph "dependency-cruiser output"{
     edge [color=black arrowhead=normal fontname=Helvetica fontsize=9]
 
     "./not-at-home" [label="not-at-home" color="red" fontcolor="red" ]
-    subgraph "cluster_/." {label="." subgraph "cluster_/./this" {label="this" subgraph "cluster_/./this/path" {label="path" subgraph "cluster_/./this/path/does" {label="does" subgraph "cluster_/./this/path/does/not" {label="not" "./this/path/does/not/exist" [label="exist" color="red" fontcolor="red" ] } } } } }
-    subgraph "cluster_/test" {label="test" subgraph "cluster_/test/cli" {label="cli" subgraph "cluster_/test/cli/fixtures" {label="fixtures" subgraph "cluster_/test/cli/fixtures/unresolvable-in-sub" {label="unresolvable-in-sub" "test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js" [label="refers-to-an-unresolvable-module.js" URL="STUFFCOMINGBEFORE/test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js"] } } } }
+    subgraph "cluster_." {label="." subgraph "cluster_./this" {label="this" subgraph "cluster_./this/path" {label="path" subgraph "cluster_./this/path/does" {label="does" subgraph "cluster_./this/path/does/not" {label="not" "./this/path/does/not/exist" [label="exist" color="red" fontcolor="red" ] } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/unresolvable-in-sub" {label="unresolvable-in-sub" "test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js" [label="refers-to-an-unresolvable-module.js" URL="STUFFCOMINGBEFORE/test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js"] } } } }
 
     "test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js" -> "./not-at-home"
     "test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js" -> "./this/path/does/not/exist"
